Use ro.benzo.version to signal mIsUpgrade
M: We use a static fingerprint that is only changed when
a new OEM build is released, so every flash shows Android
is starting instead of upgrading. This will fix that.

N: even though we dont have the dexopt sceen on N, this
is still needed to delete the correct caches, and grant/deny
specific runtime permissions like a true oem update would do

Change-Id: I0e3ed5c8f0351e48944432ae6a0c5194ddeff1fa
Signed-off-by: Jason Edson <jaysonedson@gmail.com>