ConnectivityManager is initialized with a wrong context
Symptom:
BT tethering was failed due to security exception. The error message
said it's requested by "android" package without system uid.

Root cause:
ActivityThread#setHttpProxy uses a system context to get
a ConnectivityManager instance. Unexpectedly, ConnectivityManager with
the system context is cached to the system service registry when it's
not cached yet. The wrong context is always reused when getting
ConnectivityManager.

Solution:
Use an Application context to get ConnectivityManager if it can be used.

Bug: 73572062
Change-Id: Id652c1625194dc73dcbf8dfcad4b4e2e6fbebf4e
Signed-off-by: mydongistiny <jaysonedson@gmail.com>