services: Teach PinnerService to pin parts of APKs
This change teaches PinnerService how to pin parts of APK files, with
the specific regions of interest described by a small metadata file in
the APK root directory. Another change will include a small tool that
decorates a zip file with the additional information that the new
PinnerService reads in order to determine the specific regions of that
file to pin.

This CL also fixes a few resource management bugs and enhances the
dumpsys output.

Test: it builds?
Change-Id: Ic05fd5b13abd4929af4180dcb70ac9a385feaeb1
Signed-off-by: mydongistiny <jaysonedson@gmail.com>