HeadsUpAppearanceController: allow heads-up to draw beyond bounds
On some devices with huge `rounded_corner_content_padding`, the icon of
heads-up notifications will be clipped or even completely removed from
view. This is caused by `HeadsUpStatusBarView` trying to align the
heads-up icon with where the icon would be on a normal notification
card, which would end up outside of the status bar contents view when
`rounded_corner_content_padding` is set to beyond a certain value.

This patch fixes the problem by allowing heads-up content to overflow
the normal status bar content region when it is shown. As long as the
rounded corner is not too huge, the aligning position of the icon should
not be a problem of its own.

Change-Id: I0fb8856c301a438ca87b06fcdbcf5323d9cda18c
Signed-off-by: mydongistiny <jaysonedson@gmail.com>