When data switches during voice call, make sure signal bar updates it
isDataEnabled only reflects user setting of mobile data is on or off.
isDataCapable reflects better of whether the data could be setup or not,
considering user settings and other restrictions and exceptions.

We should use isDataCapable when deciding what to show in signal bar.

When non default data SIM is in voice call, there could be an exception
that we also setup data on that SIM.

Bug: 134981077
Test: manual
Change-Id: I54fc4ca104cb451af847d7973fbffc7f0ed97c10
Merged-In: I54fc4ca104cb451af847d7973fbffc7f0ed97c10
Signed-off-by: mydongistiny <jaysonedson@gmail.com>