[hwui] fix -Wimplicit-int-float-conversion
DIRTY_MAX cannot be represented precisely in an IEEE 754 single
precision float.  Calls to DamageAccumulator::dirty() that pass
DIRTY_MAX thus result in an implicit cast that results in imprecision.
Add one to the DIRTY_MAX constant so that its value can be precisely
representing in a single precision float.

Alternatively, we can accept the imprecision via explicit casts in
multiple call sites.

Bug: 139945549
Test: mm
Change-Id: I38b2942f976ad05af2d0e45d1a9fe4214e759a02
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>